+++ invenio_config_iform/custom_fields/__init__.py
@@ -2,7 +2,7 @@
 #
 # Copyright (C) 2024 Graz University of Technology.
 #
-# invenio-config-tugraz is free software; you can redistribute it and/or
+# invenio-config-iform is free software; you can redistribute it and/or
 # modify it under the terms of the MIT License; see LICENSE file for more
 # details.