+++ invenio_config_iform/views.py
@@ -2,11 +2,11 @@
 #
 # Copyright (C) 2020-2024 Graz University of Technology.
 #
-# invenio-config-tugraz is free software; you can redistribute it and/or
+# invenio-config-iform is free software; you can redistribute it and/or
 # modify it under the terms of the MIT License; see LICENSE file for more
 # details.
 
-"""invenio module for TUGRAZ config."""
+"""Invenio module for I-Form config."""
 
 from flask import Blueprint, Flask, redirect, url_for
 from invenio_i18n import get_locale
@@ -14,11 +14,11 @@ from werkzeug.wrappers import Response as BaseResponse
 
 
 def ui_blueprint(app: Flask) -> Blueprint:
-    """Blueprint for the routes and resources provided by invenio-config-tugraz."""
-    routes = app.config.get("CONFIG_TUGRAZ_ROUTES")
+    """Blueprint for the routes and resources provided by invenio-config-iform."""
+    routes = app.config.get("CONFIG_IFORM_ROUTES")
 
     blueprint = Blueprint(
-        "invenio_config_tugraz",
+        "invenio_config_iform",
         __name__,
         template_folder="templates",
         static_folder="static",
@@ -32,12 +32,12 @@ def ui_blueprint(app: Flask) -> Blueprint:
 
 
 def guide() -> BaseResponse:
-    """TUGraz_Repository_Guide."""
+    """I-Form_Repository_Guide."""
     locale = get_locale()
     return redirect(
         url_for(
             "static",
-            filename=f"documents/TUGraz_Repository_Guide_02.1_{locale}.pdf",
+            filename=f"documents/I-Form_Repository_Guide_02.1_{locale}.pdf",
             _external=True,
         ),
     )
@@ -49,7 +49,7 @@ def terms() -> BaseResponse:
     return redirect(
         url_for(
             "static",
-            filename=f"documents/TUGraz_Repository_Terms_And_Conditions_{locale}.pdf",
+            filename=f"documents/I-Form_Repository_Terms_And_Conditions_{locale}.pdf",
             _external=True,
         ),
     )
@@ -61,7 +61,7 @@ def gdpr() -> BaseResponse:
     return redirect(
         url_for(
             "static",
-            filename=f"documents/TUGraz_Repository_General_Data_Protection_Rights_{locale}.pdf",
+            filename=f"documents/I-Form_Repository_General_Data_Protection_Rights_{locale}.pdf",
             _external=True,
         ),
     )