+++ run-tests.sh
@@ -5,7 +5,7 @@
 # Copyright (C) 2019-2020 Northwestern University.
 # Copyright (C) 2020-2024 Graz University of Technology.
 #
-# invenio-config-tugraz is free software; you can redistribute it and/or
+# invenio-config-iform is free software; you can redistribute it and/or
 # modify it under the terms of the MIT License; see LICENSE file for more
 # details.